Canvas layers!
- Added another canvas layer; now one is for the moving tetromino, and
one is for the landed ones. This will possibly give better performance
as the landed tetrominos won't have to be redrawn all the time.
- Renamed all global variables to start with 'g_' in order to try to fix
a bug. However, the bug was fixed by other means and I have yet to
decide if I want to keep this naming convention or not.
- Added some neat block borders.
- The overlays (showing 'Game Over' etc.) are yet to be reimplemented.